+++ Avionics_RX/src/RFM95W.cpp
@@ -71,7 +71,7 @@ std::string RFM95W::UDP_Receive() const
 
 }
 
-std::string RFM95W::TCP_Receive(const uint8_t &Time_Out) const
+std::string RFM95W::TCP_Receive(const uint16_t &Time_Out) const
 {
     uint8_t Buffer[m_Max_message_length];
     // isn't this just m_Max_Message_Length + 1
@@ -100,7 +100,7 @@ std::string RFM95W::TCP_Receive(const uint8_t &Time_Out) const
 
 }
 
-std::string RFM95W::UDP_Receive(const uint8_t &Time_Out) const
+std::string RFM95W::UDP_Receive(const uint16_t &Time_Out) const
 {
     uint8_t Buffer[m_Max_message_length];
     // isn't this just m_Max_Message_Length + 1
